refactor(home): criação de screenHome
